requirements.txt
streamlit
pytesseract
Pillow
openai
requests
beautifulsoup4
tweepy